Split term.rs
- Term lives in term.rs (flat)
- TermDag lives in term_dag.rs (linked)
- Value lives in value.rs (not a term)
