+++ src/core/term.rs
@@ -1,12 +1,6 @@
-use super::types;
 use crate::extension::Jet;
-use crate::util::slice_to_u64_be;
-use crate::Error;
 use crate::{cmr, extension};
-use std::collections::HashMap;
-use std::fmt;
-use std::hash::{Hash, Hasher};
-use std::rc::Rc;
+use std::hash::Hash;
 
 /// Simplicity expression node, including Bitcoin/Elements extensions
 ///
@@ -95,53 +89,6 @@ impl<Witness, Extension: extension::Jet> Term<Witness, Extension> {
     }
 }
 
-/// Simplicity expression node, including Bitcoin/Elements extensions
-///
-/// If Bitcoin/Elements support is not compiled (see `bitcoin` and
-/// `elements` features) programs using these extensions will fail to
-/// parse.
-///
-/// The structure stores the simplicity program as a directed acyclic graph(DAG).
-/// This structure is useful for creating simplicity programs recursively.
-#[derive(Clone, Eq, PartialEq, PartialOrd, Ord, Hash, Debug)]
-pub enum TermDag<Witness, Extension> {
-    Iden,
-    Unit,
-    InjL(Rc<TermDag<Witness, Extension>>),
-    InjR(Rc<TermDag<Witness, Extension>>),
-    Take(Rc<TermDag<Witness, Extension>>),
-    Drop(Rc<TermDag<Witness, Extension>>),
-    Comp(
-        Rc<TermDag<Witness, Extension>>,
-        Rc<TermDag<Witness, Extension>>,
-    ),
-    Case(
-        Rc<TermDag<Witness, Extension>>,
-        Rc<TermDag<Witness, Extension>>,
-    ),
-    AssertL(
-        Rc<TermDag<Witness, Extension>>,
-        Rc<TermDag<Witness, Extension>>,
-    ),
-    AssertR(
-        Rc<TermDag<Witness, Extension>>,
-        Rc<TermDag<Witness, Extension>>,
-    ),
-    Pair(
-        Rc<TermDag<Witness, Extension>>,
-        Rc<TermDag<Witness, Extension>>,
-    ),
-    Disconnect(
-        Rc<TermDag<Witness, Extension>>,
-        Rc<TermDag<Witness, Extension>>,
-    ),
-    Witness(Witness),
-    Fail([u8; 32], [u8; 32]),
-    Hidden(cmr::Cmr),
-    Ext(Extension),
-    Jet(extension::jets::JetsNode),
-}
-
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
 pub struct UnTypedProg<Witness, Extension>(pub Vec<Term<Witness, Extension>>);
 
@@ -159,362 +106,5 @@ impl<Witness, Extension> UnTypedProg<Witness, Extension> {
     /// Returns an iterator over the (shared) terms in the program
     pub fn iter(&self) -> impl Iterator<Item = &Term<Witness, Extension>> {
         self.0.iter()
-    }
-}
-
-impl<Witness, Extension> TermDag<Witness, Extension> {
-    /// Create a DAG representation from an untyped representation
-    pub fn from_untyped_prog(untyped_prog: UnTypedProg<Witness, Extension>) -> Rc<Self> {
-        assert!(
-            !untyped_prog.0.is_empty(),
-            "Untyped Program len must be greater than 0"
-        );
-        let mut dag_list: Vec<Rc<TermDag<_, _>>> = vec![];
-        for (index, term) in untyped_prog.0.into_iter().enumerate() {
-            let dag = match term {
-                Term::Iden => Rc::new(TermDag::Iden),
-                Term::Unit => Rc::new(TermDag::Unit),
-                Term::InjL(l) => Rc::new(TermDag::InjL(Rc::clone(&dag_list[index - l]))),
-                Term::InjR(r) => Rc::new(TermDag::InjR(Rc::clone(&dag_list[index - r]))),
-                Term::Take(l) => Rc::new(TermDag::Take(Rc::clone(&dag_list[index - l]))),
-                Term::Drop(r) => Rc::new(TermDag::Drop(Rc::clone(&dag_list[index - r]))),
-                Term::Comp(l, r) => Rc::new(TermDag::Comp(
-                    Rc::clone(&dag_list[index - l]),
-                    Rc::clone(&dag_list[index - r]),
-                )),
-                Term::Case(l, r) => Rc::new(TermDag::Case(
-                    Rc::clone(&dag_list[index - l]),
-                    Rc::clone(&dag_list[index - r]),
-                )),
-                Term::AssertL(l, r) => Rc::new(TermDag::AssertL(
-                    Rc::clone(&dag_list[index - l]),
-                    Rc::clone(&dag_list[index - r]),
-                )),
-                Term::AssertR(l, r) => Rc::new(TermDag::AssertR(
-                    Rc::clone(&dag_list[index - l]),
-                    Rc::clone(&dag_list[index - r]),
-                )),
-                Term::Pair(l, r) => Rc::new(TermDag::Pair(
-                    Rc::clone(&dag_list[index - l]),
-                    Rc::clone(&dag_list[index - r]),
-                )),
-                Term::Disconnect(l, r) => Rc::new(TermDag::Disconnect(
-                    Rc::clone(&dag_list[index - l]),
-                    Rc::clone(&dag_list[index - r]),
-                )),
-                Term::Witness(w) => Rc::new(TermDag::Witness(w)),
-                //TODO: understand how Fail works and rename `a` and `b`
-                Term::Fail(a, b) => Rc::new(TermDag::Fail(a, b)),
-                Term::Hidden(c) => Rc::new(TermDag::Hidden(c)),
-                Term::Ext(e) => Rc::new(TermDag::Ext(e)),
-                Term::Jet(j) => Rc::new(TermDag::Jet(j)),
-            };
-            dag_list.push(dag);
-        }
-        Rc::clone(dag_list.last().unwrap())
-    }
-}
-
-// A direct comparison for Rc<> results in comparison
-// of underllying inner values whereas we desire to
-// compare the referececs.
-#[derive(Debug)]
-struct RcWrapper<Witness, Extension> {
-    rc: Rc<TermDag<Witness, Extension>>,
-}
-
-impl<Witness, Extension> PartialEq for RcWrapper<Witness, Extension> {
-    fn eq(&self, other: &Self) -> bool {
-        Rc::ptr_eq(&self.rc, &other.rc)
-    }
-}
-
-impl<Witness, Extension> Eq for RcWrapper<Witness, Extension> {}
-
-impl<Witness, Extension> From<Rc<TermDag<Witness, Extension>>> for RcWrapper<Witness, Extension> {
-    fn from(dag: Rc<TermDag<Witness, Extension>>) -> Self {
-        Self { rc: dag }
-    }
-}
-
-impl<Witness, Extension> Hash for RcWrapper<Witness, Extension>
-where
-    Witness: Hash,
-    Extension: Hash,
-{
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.rc.hash(state);
-    }
-}
-
-impl<Witness, Extension> TermDag<Witness, Extension>
-where
-    Witness: Hash + Clone,
-    Extension: Hash + Clone,
-{
-    /// Convert a TermDag into into a untyped program vec.
-    pub fn into_untyped_prog(self) -> UnTypedProg<Witness, Extension> {
-        // helper function to recrusively compute the index positions
-        // of the children.
-        fn into_helper<Witness, Extension>(
-            dag: Rc<TermDag<Witness, Extension>>,
-            index_map: &mut HashMap<RcWrapper<Witness, Extension>, usize>,
-            prog: &mut Vec<Term<Witness, Extension>>,
-        ) -> usize
-        where
-            RcWrapper<Witness, Extension>: Hash,
-            Witness: Clone,
-            Extension: Clone,
-        {
-            // insert one child into prog
-            macro_rules! insert_one_child {
-                ($term: expr, $l : expr) => {
-                    match index_map.get(&RcWrapper::from(Rc::clone($l))) {
-                        Some(ind) => prog.push($term(prog.len() - ind)),
-                        None => {
-                            let ind = into_helper(Rc::clone($l), index_map, prog);
-                            prog.push($term(prog.len() - ind));
-                        }
-                    }
-                };
-            }
-
-            // insert two children into prog
-            macro_rules! insert_two_child {
-                ($term: expr, $l : expr, $r: expr) => {{
-                    let l_ind = match index_map.get(&RcWrapper::from(Rc::clone($l))) {
-                        Some(ind) => *ind,
-                        None => into_helper(Rc::clone($l), index_map, prog),
-                    };
-                    let r_ind = match index_map.get(&RcWrapper::from(Rc::clone($r))) {
-                        Some(ind) => *ind,
-                        None => into_helper(Rc::clone($r), index_map, prog),
-                    };
-                    prog.push($term(prog.len() - l_ind, prog.len() - r_ind));
-                }};
-            }
-
-            if index_map.contains_key(&RcWrapper::from(Rc::clone(&dag))) {
-                return *index_map.get(&RcWrapper::from(Rc::clone(&dag))).unwrap();
-            }
-            match dag.as_ref() {
-                TermDag::Unit => prog.push(Term::Unit),
-                TermDag::Iden => prog.push(Term::Iden),
-                TermDag::InjL(l) => insert_one_child!(Term::InjL, l),
-                TermDag::InjR(r) => insert_one_child!(Term::InjR, r),
-                TermDag::Take(l) => insert_one_child!(Term::Take, l),
-                TermDag::Drop(r) => insert_one_child!(Term::Drop, r),
-                TermDag::Comp(l, r) => insert_two_child!(Term::Comp, l, r),
-                TermDag::Case(l, r) => insert_two_child!(Term::Case, l, r),
-                TermDag::AssertL(l, r) => insert_two_child!(Term::AssertL, l, r),
-                TermDag::AssertR(l, r) => insert_two_child!(Term::AssertR, l, r),
-                TermDag::Pair(l, r) => insert_two_child!(Term::Pair, l, r),
-                TermDag::Disconnect(l, r) => insert_two_child!(Term::Disconnect, l, r),
-                TermDag::Witness(ref w) => prog.push(Term::Witness(w.clone())),
-                TermDag::Fail(a, b) => prog.push(Term::Fail(*a, *b)),
-                TermDag::Hidden(cmr) => prog.push(Term::Hidden(*cmr)),
-                TermDag::Ext(ref e) => prog.push(Term::Ext(e.clone())),
-                TermDag::Jet(ref j) => prog.push(Term::Jet(*j)),
-            }
-            // insert the current node remembering it's index for reusing
-            index_map.insert(RcWrapper::from(dag), prog.len() - 1);
-            prog.len() - 1
-        }
-
-        let mut prog = vec![];
-        let mut index_map = HashMap::new();
-        let _len = into_helper(Rc::new(self), &mut index_map, &mut prog);
-
-        UnTypedProg(prog)
-    }
-}
-
-#[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
-pub enum Value {
-    Unit,
-    SumL(Box<Value>),
-    SumR(Box<Value>),
-    Prod(Box<Value>, Box<Value>),
-}
-
-impl Value {
-    #![allow(clippy::len_without_is_empty)]
-    /// The length, in bits, of the value when encoded in the Bit Machine
-    pub fn len(&self) -> usize {
-        match *self {
-            Value::Unit => 0,
-            Value::SumL(ref s) => 1 + s.len(),
-            Value::SumR(ref s) => 1 + s.len(),
-            Value::Prod(ref s, ref t) => s.len() + t.len(),
-        }
-    }
-
-    /// Encode a single bit as a value. Will panic if the input is out of range
-    pub fn u1(n: u8) -> Value {
-        match n {
-            0 => Value::SumL(Box::new(Value::Unit)),
-            1 => Value::SumR(Box::new(Value::Unit)),
-            x => panic!("{} out of range for Value::u1", x),
-        }
-    }
-
-    /// Encode a two-bit number as a value. Will panic if the input is out of range
-    pub fn u2(n: u8) -> Value {
-        let b0 = (n & 2) / 2;
-        let b1 = n & 1;
-        if n > 3 {
-            panic!("{} out of range for Value::u2", n);
-        }
-        Value::Prod(Box::new(Value::u1(b0)), Box::new(Value::u1(b1)))
-    }
-
-    /// Encode a four-bit number as a value. Will panic if the input is out of range
-    pub fn u4(n: u8) -> Value {
-        let w0 = (n & 12) / 4;
-        let w1 = n & 3;
-        if n > 15 {
-            panic!("{} out of range for Value::u4", n);
-        }
-        Value::Prod(Box::new(Value::u2(w0)), Box::new(Value::u2(w1)))
-    }
-
-    /// Encode an eight-bit number as a value
-    pub fn u8(n: u8) -> Value {
-        let w0 = n >> 4;
-        let w1 = n & 0xf;
-        Value::Prod(Box::new(Value::u4(w0)), Box::new(Value::u4(w1)))
-    }
-
-    /// Encode a 16-bit number as a value
-    pub fn u16(n: u16) -> Value {
-        let w0 = (n >> 8) as u8;
-        let w1 = (n & 0xff) as u8;
-        Value::Prod(Box::new(Value::u8(w0)), Box::new(Value::u8(w1)))
-    }
-
-    /// Encode a 32-bit number as a value
-    pub fn u32(n: u32) -> Value {
-        let w0 = (n >> 16) as u16;
-        let w1 = (n & 0xffff) as u16;
-        Value::Prod(Box::new(Value::u16(w0)), Box::new(Value::u16(w1)))
-    }
-
-    /// Encode a 32-bit number as a value
-    pub fn u64(n: u64) -> Value {
-        let w0 = (n >> 32) as u32;
-        let w1 = (n & 0xffff_ffff) as u32;
-        Value::Prod(Box::new(Value::u32(w0)), Box::new(Value::u32(w1)))
-    }
-
-    /// Encode a 32 byte number into value. Useful for encoding 32 pubkeys/hashes
-    pub fn u256_from_slice(v: &[u8]) -> Value {
-        assert!(v.len() == 32);
-        Value::Prod(
-            Box::new(Value::Prod(
-                Box::new(Value::u64(slice_to_u64_be(&v[0..8]))),
-                Box::new(Value::u64(slice_to_u64_be(&v[8..16]))),
-            )),
-            Box::new(Value::Prod(
-                Box::new(Value::u64(slice_to_u64_be(&v[16..24]))),
-                Box::new(Value::u64(slice_to_u64_be(&v[24..32]))),
-            )),
-        )
-    }
-
-    /// Encode a 64(pair(32, 32)) byte number into value.
-    /// Useful for encoding 64 byte signatures
-    pub fn u512_from_slice(v: &[u8]) -> Value {
-        assert!(v.len() == 64);
-        Value::Prod(
-            Box::new(Value::u256_from_slice(&v[0..32])),
-            Box::new(Value::u256_from_slice(&v[32..64])),
-        )
-    }
-
-    /// Convert the value to a byte array.
-    pub fn into_bits(self) -> Vec<bool> {
-        let mut ret = vec![];
-        fn into_bit_helper(v: Value, ret: &mut Vec<bool>) {
-            match v {
-                Value::Unit => {}
-                Value::SumL(l) => {
-                    ret.push(false);
-                    into_bit_helper(*l, ret);
-                }
-                Value::SumR(r) => {
-                    ret.push(true);
-                    into_bit_helper(*r, ret);
-                }
-                Value::Prod(l, r) => {
-                    into_bit_helper(*l, ret);
-                    into_bit_helper(*r, ret);
-                }
-            }
-        }
-        into_bit_helper(self, &mut ret);
-        ret
-    }
-
-    /// Convenience constructor for a left sum of a value
-    pub fn sum_l(a: Value) -> Value {
-        Value::SumL(Box::new(a))
-    }
-
-    /// Convenience constructor for a right sum of a value
-    pub fn sum_r(a: Value) -> Value {
-        Value::SumR(Box::new(a))
-    }
-
-    /// Convenience constructor for product of two values
-    pub fn prod(a: Value, b: Value) -> Value {
-        Value::Prod(Box::new(a), Box::new(b))
-    }
-}
-
-impl fmt::Display for Value {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            Value::Unit => f.write_str("ε"),
-            Value::SumL(ref sub) => {
-                f.write_str("0")?;
-                if **sub != Value::Unit {
-                    write!(f, "{}", sub)
-                } else {
-                    Ok(())
-                }
-            }
-            Value::SumR(ref sub) => {
-                f.write_str("1")?;
-                if **sub != Value::Unit {
-                    write!(f, "{}", sub)
-                } else {
-                    Ok(())
-                }
-            }
-            Value::Prod(ref l, ref r) => {
-                write!(f, "({},", l)?;
-                write!(f, "{})", r)
-            }
-        }
-    }
-}
-
-impl Value {
-    pub fn from_bits_and_type<Bits: Iterator<Item = bool>>(
-        bits: &mut Bits,
-        ty: &types::FinalType,
-    ) -> Result<Value, Error> {
-        match ty.ty {
-            types::FinalTypeInner::Unit => Ok(Value::Unit),
-            types::FinalTypeInner::Sum(ref l, ref r) => match bits.next() {
-                Some(false) => Ok(Value::SumL(Box::new(Value::from_bits_and_type(bits, &*l)?))),
-                Some(true) => Ok(Value::SumR(Box::new(Value::from_bits_and_type(bits, &*r)?))),
-                None => Err(Error::EndOfStream),
-            },
-            types::FinalTypeInner::Product(ref l, ref r) => Ok(Value::Prod(
-                Box::new(Value::from_bits_and_type(&mut *bits, &*l)?),
-                Box::new(Value::from_bits_and_type(bits, &*r)?),
-            )),
-        }
     }
 }
